댓글 재귀적 삭제 기능 테스트 및 API 구현

+++ service/comment/src/main/kotlin/kuke/board/comment/repository/CommentRepository.kt
@@ -23,5 +23,5 @@ interface CommentRepository: JpaRepository<Comment, Long> {
         @Param("articleId") articleId: Long,
         @Param("parentCommentId") parentCommentId: Long,
         @Param("limit") limit: Long
-    ): Int
+    ): Long
 }
